Started adding functions for creating function lists

+++ src/Sensors/I2C/Loom_K30/Loom_K30.h
@@ -43,7 +43,7 @@ class Loom_K30 : public I2CDevice{
         /**
          * Get the current CO2 levels collected by the sensor
          */ 
-        int getCO2() { return CO2Levels; };
+        float getCO2() { return (float)CO2Levels; };
 
     private:
         Manager* manInst;                                               // Instance of the manager
